AISettingsクラスの追加とSettingsクラスの変更
AI翻訳アプリケーションの設定を表す`AISettings`クラスを新たに追加しました。このクラスには、認証トークン、チャットエンドポイントURL、モデルID、チャンクサイズ、最大同時リクエスト数のプロパティが含まれています。また、`ToString`メソッドをオーバーライドし、トークンの一部をマスクして表示する機能を追加しました。

一方、`Settings`クラスからはAI関連のプロパティが削除され、ソースドキュメントのパスや言語、出力ディレクトリのパス、ログレベルに関連するプロパティが残りました。`ToString`メソッドも更新され、AI関連のプロパティが表示されないように変更されています。

さらに、`AISettingsTest.cs`に`AISettings`クラスのテストを追加し、`ToString`メソッドの動作確認を行うテストケースを作成しました。`SettingsTest.cs`では、AI関連のプロパティに関するテストが削除され、テストがソースドキュメントや出力に関連するプロパティに焦点を当てるようになりました。
