+++ src/TsunaCan.XmlDocumentationTranslator.Core/Settings.cs
@@ -9,11 +9,6 @@ namespace TsunaCan.XmlDocumentationTranslator;
 public class Settings
 {
     /// <summary>
-    ///  Gets or sets the authentication token.
-    /// </summary>
-    public required string Token { get; set; }
-
-    /// <summary>
     ///  Gets or sets the path to the IntelliSense document.
     /// </summary>
     public required string SourceDocumentPath { get; set; }
@@ -38,42 +33,14 @@ public class Settings
     /// </summary>
     public LogLevel LogLevel { get; set; }
 
-    /// <summary>
-    ///  Gets or sets the chat endpoint URL.
-    /// </summary>
-    public required Uri ChatEndPointUrl { get; set; }
-
-    /// <summary>
-    ///  Gets or sets the model ID to be used.
-    /// </summary>
-    public required string ModelId { get; set; }
-
-    /// <summary>
-    ///  Gets or sets the chunk size for the translation process.
-    /// </summary>
-    public required int ChunkSize { get; set; } = 1000;
-
-    /// <summary>
-    ///  Gets or sets the maximum number of concurrent requests for AI translation.
-    /// </summary>
-    public int MaxConcurrentRequests { get; set; } = 4;
-
     /// <inheritdoc/>
     public override string ToString()
     {
-        string tokenDisplay = this.Token.Length > 10
-            ? $"{new string('*', this.Token.Length - 5)}{this.Token[^5..]}"
-            : new string('*', this.Token.Length);
-        return $"{nameof(this.Token)}: {tokenDisplay}, " +
-            $"{nameof(this.SourceDocumentPath)}: {this.SourceDocumentPath}, " +
+        return $"{nameof(this.SourceDocumentPath)}: {this.SourceDocumentPath}, " +
             $"{nameof(this.SourceDocumentLanguage)}: {this.SourceDocumentLanguage}, " +
             $"{nameof(this.OutputDirectoryPath)}: {this.OutputDirectoryPath}, " +
             $"{nameof(this.OutputFileLanguages)}: {this.OutputFileLanguagesString()}, " +
-            $"{nameof(this.LogLevel)}: {this.LogLevel}, " +
-            $"{nameof(this.ChatEndPointUrl)}: {this.ChatEndPointUrl}, " +
-            $"{nameof(this.ModelId)}: {this.ModelId}, " +
-            $"{nameof(this.ChunkSize)}: {this.ChunkSize}, " +
-            $"{nameof(this.MaxConcurrentRequests)}: {this.MaxConcurrentRequests}";
+            $"{nameof(this.LogLevel)}: {this.LogLevel}";
     }
 
     private string OutputFileLanguagesString()
